Checkpointing improvement on code quality.
